Better support for standalone expresssions
