+++ cli/cmd_slq.go
@@ -5,11 +5,11 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/neilotoole/sq/libsq/core/lg/lga"
+
 	"github.com/neilotoole/sq/cli/output/format"
 
 	"github.com/neilotoole/sq/cli/run"
-
-	"github.com/neilotoole/sq/libsq/core/lg/lga"
 
 	"github.com/neilotoole/sq/cli/flag"
 	"github.com/neilotoole/sq/drivers/csv"
@@ -144,9 +144,9 @@ func execSLQ(cmd *cobra.Command, args []string) error {
 func execSLQInsert(ctx context.Context, ru *run.Run, mArgs map[string]string,
 	destSrc *source.Source, destTbl string,
 ) error {
-	args, coll, dbases := ru.Args, ru.Config.Collection, ru.Databases
+	qc := run.NewQueryContext(ru, mArgs)
 
-	slq, err := preprocessUserSLQ(ctx, ru, args)
+	slq, err := preprocessUserSLQ(ctx, ru, ru.Args)
 	if err != nil {
 		return err
 	}
@@ -154,7 +154,7 @@ func execSLQInsert(ctx context.Context, ru *run.Run, mArgs map[string]string,
 	ctx, cancelFn := context.WithCancel(ctx)
 	defer cancelFn()
 
-	destDB, err := dbases.Open(ctx, destSrc)
+	destDB, err := ru.Databases.Open(ctx, destSrc)
 	if err != nil {
 		return err
 	}
@@ -170,14 +170,6 @@ func execSLQInsert(ctx context.Context, ru *run.Run, mArgs map[string]string,
 		driver.OptTuningRecChanSize.Get(destSrc.Options),
 		libsq.DBWriterCreateTableIfNotExistsHook(destTbl),
 	)
-
-	qc := &libsq.QueryContext{
-		Collection:      coll,
-		DBOpener:        ru.Databases,
-		JoinDBOpener:    ru.Databases,
-		ScratchDBOpener: ru.Databases,
-		Args:            mArgs,
-	}
 
 	execErr := libsq.ExecuteSLQ(ctx, qc, slq, inserter)
 	affected, waitErr := inserter.Wait() // Wait for the writer to finish processing
@@ -195,16 +187,11 @@ func execSLQInsert(ctx context.Context, ru *run.Run, mArgs map[string]string,
 
 // execSLQPrint executes the SLQ query, and prints output to writer.
 func execSLQPrint(ctx context.Context, ru *run.Run, mArgs map[string]string) error {
+	qc := run.NewQueryContext(ru, mArgs)
+
 	slq, err := preprocessUserSLQ(ctx, ru, ru.Args)
 	if err != nil {
 		return err
-	}
-
-	qc := &libsq.QueryContext{
-		Collection:   ru.Config.Collection,
-		DBOpener:     ru.Databases,
-		JoinDBOpener: ru.Databases,
-		Args:         mArgs,
 	}
 
 	recw := output.NewRecordWriterAdapter(ctx, ru.Writers.Record)
@@ -222,7 +209,7 @@ func execSLQPrint(ctx context.Context, ru *run.Run, mArgs map[string]string) err
 // function is something of a hangover from the early days of
 // sq and may need to be rethought.
 //
-// 1. If there's piped input but no query args, the first table
+// If there's piped input but no query args, the first table
 // from the pipe source becomes the query. Invoked like this:
 //
 //	$ cat something.csv | sq
@@ -234,13 +221,6 @@ func execSLQPrint(ctx context.Context, ru *run.Run, mArgs map[string]string) err
 // For non-monotable sources, the first table is used:
 //
 //	$ cat something.xlsx | sq @stdin.sheet1
-//
-// 2. If the query doesn't contain a source selector segment
-// starting with @HANDLE, the active src handle is prepended
-// to the query. This allows a query where the first selector
-// segment is the table name.
-//
-//	$ sq '.person'  -->  $ sq '@active.person'
 func preprocessUserSLQ(ctx context.Context, ru *run.Run, args []string) (string, error) {
 	log, reg, dbases, coll := lg.FromContext(ctx), ru.DriverRegistry, ru.Databases, ru.Config.Collection
 	activeSrc := coll.Active()
@@ -249,7 +229,7 @@ func preprocessUserSLQ(ctx context.Context, ru *run.Run, args []string) (string,
 		// Special handling for the case where no args are supplied
 		// but sq is receiving pipe input. Let's say the user does this:
 		//
-		//  $ cat something.csv | sq  # query becomes ".stdin.data"
+		//  $ cat something.csv | sq  # query becomes "@stdin.data"
 		if activeSrc == nil {
 			// Piped input would result in an active @stdin src. We don't
 			// have that; we don't have any active src.
@@ -293,11 +273,13 @@ func preprocessUserSLQ(ctx context.Context, ru *run.Run, args []string) (string,
 				return "", errz.New(msgSrcEmptyTableName)
 			}
 
-			log.Debug("Using first table name from document source metadata as table selector: ", tblName)
+			log.Debug("Using first table name from document source metadata as table selector",
+				lga.Src, activeSrc, lga.Table, tblName)
 		}
 
 		selector := source.StdinHandle + "." + tblName
-		log.Debug("Added selector to argument-less piped query: ", selector)
+		log.Debug("Added selector to argument-less piped query",
+			lga.Handle, source.StdinHandle, "selector", selector)
 
 		return selector, nil
 	}
@@ -334,18 +316,7 @@ func preprocessUserSLQ(ctx context.Context, ru *run.Run, args []string) (string,
 		return query, nil
 	}
 
-	// The query doesn't start with a handle selector; let's prepend
-	// a handle selector segment.
-	if activeSrc == nil {
-		return "", errz.New("no data source provided, and no active data source")
-	}
-
 	query := strings.Join(args, " ")
-	query = fmt.Sprintf("%s | %s", activeSrc.Handle, query)
-
-	log.Debug("The query didn't start with @handle, so the active src was prepended",
-		lga.Query, query)
-
 	return query, nil
 }
