+++ libsq/libsq.go
@@ -31,10 +31,10 @@ type QueryContext struct {
 	// DBOpener is used to open databases.
 	DBOpener driver.DatabaseOpener
 
-	// JoinDBOpener is used to open the joindb (if needed).
+	// JoinDBOpener is used to open the joindb.
 	JoinDBOpener driver.JoinDatabaseOpener
 
-	// ScratchDBOpener is used to open the scratchdb (if needed).
+	// ScratchDBOpener is used to open the scratchdb.
 	ScratchDBOpener driver.ScratchDatabaseOpener
 
 	// Args defines variables that are substituted into the query.
